move computeDisplayDates and isExpiring to access-utils
